requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cssselect2==0.8.0
distro==1.9.0
Flask==3.1.1
fonttools==4.58.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.79.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
pyphen==0.17.2
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.1.0
requests==2.32.3
selenium==4.32.0
sniffio==1.3.1
sortedcontainers==2.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
weasyprint==65.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zopfli==0.2.3.post1
